Fixed delete log not working.
