+++ .gitignore
@@ -7,5 +7,5 @@ flask/
 #Database Migrations
 migrations/
 
-#Upload Folder for logs
-app/uploads/*
+#Upload Folder for logs. Keep actual logs out, but retain the directory
+app/uploads/logs/*.bin
